requirements.txt
django
psycopg2-binary
pytest
pytest-django>=4.5
djangorestframework>=3.14.0
drf-yasg>=1.21.6    
pyyaml
django-cors-headers>=4.3.1
django-filter>=23.0